Make count/has modes slightly more strict
You could previously `count` boxes in random data and there was a good
chance you'd get a number back.  It now tries to verify that the file at
least consists of complete boxes.  `has` mode is still very lax, but at
least checks that the box it finds is followed by another valid header
or EOF.